Resources for main dashboard - fix db units (#678)
* wip

* mod + sum

* added docker and dev resources

* add connection status

* model

* small fixes

* desired and actual pods

* added max and current usage

* added proxy system info for dev env

* small fix

* fixed db stats for docker env

* fix proxy ports for dev env

* fix address for docker env

* add logs to k8s test

* go mod sum

* k8s components pod and container fixes

* namespace fix

* fix percentage

* try with string

* check limits

* typo

* try parse float

* value check

* check mem & cpu k8s

* typo

* log error fix

* check http get k8s api

* https

* try get with cacert and bearer

* try pvc size and usage

* check pvc list

* PersistentVolumeClaims

* check with api request

* check storage k8s

* add ctx to exec

* throughput

* comment out unneccessary code

* generic create system streams

* remove comments + unneccessary code

* remove unneccessary code and packages

* remove comment

* removed duplicated code + fixed bugs

* fix db units

* fix bracets

+++ server/memphis_handlers_monitoring.go
@@ -121,7 +121,7 @@ func (mh MonitoringHandler) GetSystemComponents() ([]models.SystemComponents, er
 				storageComp = models.CompStats{
 					Total:      storage_size,
 					Current:    int64(v.JetStream.Stats.Store),
-					Percentage: int(math.Ceil(float64(v.JetStream.Stats.Store)/float64(storage_size)) * 100),
+					Percentage: int(math.Ceil((float64(v.JetStream.Stats.Store) / float64(storage_size)) * 100)),
 				}
 			}
 			brokerComponents = append(brokerComponents, models.SysComponent{
@@ -1489,9 +1489,9 @@ func getDbStorageSize() (int64, int64, error) {
 		return 0, 0, err
 	}
 
-	dbStorageSize := sbStats.Lookup("dataSize").Int64() + sbStats.Lookup("indexSize").Int64()
-	totalSize := sbStats.Lookup("fsTotalSize").Int64()
-	return dbStorageSize, totalSize, nil
+	dbStorageSize := sbStats.Lookup("dataSize").Double() + sbStats.Lookup("indexSize").Double()
+	totalSize := sbStats.Lookup("fsTotalSize").Double()
+	return int64(dbStorageSize), int64(totalSize), nil
 }
 
 func getUnixStorageSize() (int64, error) {
